requirements.txt
alembic==1.12.0
blinker==1.6.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Compress==1.14
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mysqlclient==2.2.0
packaging==23.1
pip==23.2.1
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
setuptools==65.5.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7